# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestionar múltiples sesiones de Telegram via MTProto",
        "title": "Telegram API",
        "contact": {},
        "version": "0.3.0"
    },
    "host": "localhost:7789",
    "basePath": "/api/v1",
    "paths": {
...
        "/sessions/{id}/messages/audio": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Envía archivo de audio",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Enviar audio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Audio",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.AudioMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.MessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{id}/messages/bulk": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Envía mensaje de texto a múltiples destinatarios con delay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Envío masivo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Mensaje masivo",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.BulkTextRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.MessageResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{id}/messages/file": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Envía archivo/documento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Enviar documento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Archivo",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.FileMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.MessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{id}/messages/photo": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Envía imagen con caption opcional",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Enviar foto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Foto",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PhotoMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.MessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{id}/messages/text": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Envía mensaje de texto simple",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Enviar texto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Mensaje de texto",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.TextMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.MessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{id}/messages/video": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Envía video con caption opcional",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Enviar video",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Video",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.VideoMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.MessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.AudioMessageRequest": {
            "description": "Mensaje con audio",
            "type": "object",
            "required": [
                "audio_url",
                "to"
            ],
            "properties": {
                "audio_url": {
                    "type": "string",
                    "example": "https://example.com/audio.mp3"
                },
                "caption": {
                    "type": "string",
                    "example": "Escucha esto"
                },
                "to": {
                    "type": "string",
                    "example": "@username"
                }
            }
        },
...
        "domain.BulkTextRequest": {
            "description": "Envío masivo de texto a múltiples destinatarios",
            "type": "object",
            "required": [
                "recipients",
                "text"
            ],
            "properties": {
                "delay_ms": {
                    "type": "integer",
                    "example": 3000
                },
                "recipients": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "@user1",
                        "@user2",
                        "+573001234567"
                    ]
                },
                "text": {
                    "type": "string",
                    "example": "Mensaje para todos"
                }
            }
        },
...
        "domain.FileMessageRequest": {
            "description": "Mensaje con archivo/documento",
            "type": "object",
            "required": [
                "file_url",
                "to"
            ],
            "properties": {
                "caption": {
                    "type": "string",
                    "example": "Documento adjunto"
                },
                "file_url": {
                    "type": "string",
                    "example": "https://example.com/doc.pdf"
                },
                "to": {
                    "type": "string",
                    "example": "@username"
                }
            }
        },
...
        "domain.MessageResponse": {
            "description": "Respuesta de envío de mensaje",
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "message": {
                    "type": "string",
                    "example": "Mensaje en cola"
                },
                "send_at": {
                    "type": "string"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.MessageStatus"
                        }
                    ],
                    "example": "pending"
                }
            }
        },
        "domain.MessageStatus": {
            "type": "string",
            "enum": [
                "pending",
                "scheduled",
                "sending",
                "sent",
                "failed"
            ],
            "x-enum-varnames": [
                "MessageStatusPending",
                "MessageStatusScheduled",
                "MessageStatusSending",
                "MessageStatusSent",
                "MessageStatusFailed"
            ]
        },
...
        "domain.PhotoMessageRequest": {
            "description": "Mensaje con foto",
            "type": "object",
            "required": [
                "photo_url",
                "to"
            ],
            "properties": {
                "caption": {
                    "type": "string",
                    "example": "Mira esta imagen"
                },
                "photo_url": {
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                },
                "to": {
                    "type": "string",
                    "example": "@username"
                }
            }
        },
...
        "domain.TextMessageRequest": {
            "description": "Mensaje de texto simple",
            "type": "object",
            "required": [
                "text",
                "to"
            ],
            "properties": {
                "text": {
                    "type": "string",
                    "example": "Hola desde la API!"
                },
                "to": {
                    "type": "string",
                    "example": "@username o +573001234567"
                }
            }
        },
...
        "domain.VideoMessageRequest": {
            "description": "Mensaje con video",
            "type": "object",
            "required": [
                "to",
                "video_url"
            ],
            "properties": {
                "caption": {
                    "type": "string",
                    "example": "Video interesante"
                },
                "to": {
                    "type": "string",
                    "example": "@username"
                },
                "video_url": {
                    "type": "string",
                    "example": "https://example.com/video.mp4"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.MetaResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "handler.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/handler.ErrorResponse"
                },
                "meta": {
                    "$ref": "#/definitions/handler.MetaResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}